+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to
 - Support for debian12 (Debian bookworm) in OS images sources YAML file.
 - Introduce fhpc_addresses extra variable, a hash with containers as keys and
   the list of IP addresses as values.
-- cli: Support for tags to filter deployed configuration tasks.
+- cli:
+  - Support for tags to filter deployed configuration tasks.
+  - Report cluster status in JSON format with --json option.
 - conf:
   - Optional support of Rackslab developement deb repositories, disabled by
     default.